backend/requirements.txt
Flask>=3.1.0
Flask-CORS>=6.0.0
opencv-python>=4.12.0
numpy>=2.2.0
Pillow>=11.3.0
requests>=2.31.0
python-dotenv>=1.1.0
gunicorn>=23.0.0
tensorflow>=2.15.0
matplotlib>=3.7.0
scikit-learn>=1.3.0 